refactored each type handler into its own class... some todos left thou
