+++ kompendium-oas/src/main/kotlin/io/bkbn/kompendium/oas/schema/EmptySchema.kt
@@ -1,0 +1,6 @@
+package io.bkbn.kompendium.oas.schema
+
+/**
+ * Placeholder schema to be stored in cache to prevent stack overflow of recursive types
+ */
+object EmptySchema : ComponentSchema
